+++ coursework/src/sample/models/Movie.java
@@ -7,26 +7,26 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo;
 @JsonTypeInfo(use=JsonTypeInfo.Id.CLASS, include=JsonTypeInfo.As.PROPERTY, property="@class")
 public class Movie {
     private String name;
-    private int rating;
+    private Double rating;
     public Integer id = null;
 
     public Movie () {};
 
-    public Movie(String name, int rating) {
+    public Movie(String name, Double rating) {
         this.name = name;
         this.rating = rating;
     }
 
     @Override
     public String toString() {
-        return String.format("Кино: \"%s\" с рейтингом %d", name, rating);
+        return String.format("Кино: \"%s\" с рейтингом %.2f", name, rating);
     }
 
     public String getName() {
         return name;
     }
 
-    public int getRating() {
+    public Double getRating() {
         return rating;
     }
 
@@ -34,7 +34,7 @@ public class Movie {
         this.name = name;
     }
 
-    public void setRating(int rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
